feature-listの構造変更(UI以外)
