--- src/feature/rentalHouse/component/admin-create/type/index.ts
@@ -1,6 +1,0 @@
-import { ReRentalHouse } from "../../../type/rentalHouse";
-
-// type
-export type CreateRentalHouse = 
-  Omit<ReRentalHouse, 'id' | 'mansion'>;
-
