+++ src/feature/rentalHouse/components/admin-create/type/index.ts
@@ -1,0 +1,7 @@
+import { RentalHouseModel } from "@/feature/rentalHouse/models/rentalHouse.model";
+
+// type
+export type CreateRentalHouse = 
+  Omit<RentalHouseModel, 'id' | 'mansion' | 'structure_type'> 
+  & {structure_type_id: number};
+
