+++ src/feature/rentalHouse/components/admin-list/components/AdminRentalHouseCard/index.tsx
