+++ src/feature/rentalHouse/components/list/page/index.tsx
@@ -1,8 +1,9 @@
 import { RentalHouseSearchBox } from "../components/RentalHouseSearchBox";
 import { RentalHouseList } from "../components/RentalHouseList";
+import { RentalHouseModel } from "@/feature/rentalHouse/models/rentalHouse.model";
 
 type Props = {
-  rentalHouses: any
+  rentalHouses: RentalHouseModel[]
 }
 
 export const SearchableRentalHouseList = ({ rentalHouses }: Props): JSX.Element => {
