+++ src/pages/admin/rentalHouse/addRentalHouse.tsx
@@ -1,5 +1,5 @@
 import { AdminLayout } from "@/components/layouts/Layout/AdminLayout";
-import { AddRentalHouse } from "@/feature/rentalHouse/component/admin-create/page";
+import { AddRentalHouse } from "@/feature/rentalHouse/components/admin-create/page";
 import { ReactElement } from "react";
 
 const AddRentalHousePage = (): JSX.Element => <AddRentalHouse />;
