+++ src/pages/admin/rentalHouse/index.tsx
@@ -1,5 +1,5 @@
 import { AdminLayout } from "@/components/layouts/Layout/AdminLayout";
-import { AdminRentalHouseList } from "@/feature/rentalHouse/component/admin-list/page";
+import { AdminRentalHouseList } from "@/feature/rentalHouse/components/admin-list/page";
 import { ReactElement } from "react";
 
 const RentalHouseListPage = (): JSX.Element => <AdminRentalHouseList />;
